Completing Task 2:
Styles on the anchors on the edit page aren't ideal, but wanted to wrap up the second readme today.

I really wanted to play with the transform property in inputs, so I updated the Angular version by a minor bump.

Added functionality to retrieve a task and to update a task. Added tests for those scenarios.