Completing Task 5:
Kind of a rush job, went back and added the delete confirmation to the todo-list (that guy is pretty thick now, I would separate some of that template and not even have the logic in the component).

The edit was pretty simple, just added the type for the form there... it's not ever present. If I have time this weekend I will restructure it to account for task 5 specifically instead of being available everywhere.